Unify tab icons with shared animated component and haptic feedback
